requirements.txt
# ── LangChain core ───────────────────────────────
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# ── LLM backends ─────────────────────────────────
langchain-ollama>=0.2.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0

# ── Embeddings ────────────────────────────────────
langchain-huggingface>=0.1.0
sentence-transformers>=3.0.0
huggingface-hub>=0.24.0
transformers>=4.44.0

# ── Vector store ──────────────────────────────────
# langchain-chroma replaces langchain_community.vectorstores.Chroma (LangChain 0.3+)
chromadb>=0.5.0
langchain-chroma>=0.1.0

# ── LangGraph ─────────────────────────────────────
langgraph>=0.2.0

# ── Document loaders ──────────────────────────────
pypdf>=4.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# ── Atlassian APIs ────────────────────────────────
atlassian-python-api>=3.41.0

# ── Utilities ─────────────────────────────────────
python-dotenv>=1.0.0
requests>=2.31.0

# ── Web UI ────────────────────────────────────────
streamlit>=1.35.0

# ── Scheduler ─────────────────────────────────────
APScheduler>=3.10.0

# ── Testing ───────────────────────────────────────
pytest>=8.0.0
pytest-mock>=3.14.0
